50 fab numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -2,12 +2,12 @@
 
 /**
  * main - print the first  50 fibonacci numbers, starting with 1 and 2
- * Return: 0 (Success)
+ * Return: Always 0
  */
 int main(void)
 {
 	int count;
-	unsigned int fib1 = 0, fib2 = 1, sum;
+	unsigned long fib1 = 0, fib2 = 1, sum;
 
 	for (count = 0; count < 50; count++)
 	{
